third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-events.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Events, implicit button
  assert_array_equals: after showing, events from test_driver.click expected property 6 to be "focusout on select at select" but got "pointerup on select at select" (expected array ["pointerdown on select at select", "pointerdown on select at wrapper", "mousedown on select at select", "mousedown on select at wrapper", "focusin on select at select", "focusin on select at wrapper", "focusout on select at select", "focusout on select at wrapper", "focusin on option1 at select", "focusin on option1 at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper", "click on select at select", "click on select at wrapper"] got ["pointerdown on select at select", "pointerdown on select at wrapper", "mousedown on select at select", "mousedown on select at wrapper", "focusin on select at select", "focusin on select at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper", "click on select at select", "click on select at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper"])
[FAIL] Events, explicit button
  assert_array_equals: after showing, events from test_driver.click expected property 6 to be "focusout on select at select" but got "pointerup on select at select" (expected array ["pointerdown on select at select", "pointerdown on select at wrapper", "mousedown on select at select", "mousedown on select at wrapper", "focusin on select at select", "focusin on select at wrapper", "focusout on select at select", "focusout on select at wrapper", "focusin on option1 at select", "focusin on option1 at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper", "click on select at select", "click on select at wrapper"] got ["pointerdown on select at select", "pointerdown on select at wrapper", "mousedown on select at select", "mousedown on select at wrapper", "focusin on select at select", "focusin on select at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper", "click on select at select", "click on select at wrapper", "pointerup on select at select", "pointerup on select at wrapper", "mouseup on select at select", "mouseup on select at wrapper"])
Harness: the test ran to completion.
